- add models schemas & resources - add all GET endpoints

+++ models.py
@@ -28,7 +28,7 @@ class Article(db.Model):
 
 	CodeArticle = Column(Integer, primary_key=True)
 	Designation = Column(String(50), default=None)
-	Poids = Column(Numeric, default=0.0000)
+	Poids = Column(Float, default=0.0000)
 	NbreDePoints = Column(Integer, default=0)
 
 class Commande(db.Model):
